Address automated review feedback on the modal-consolidation PR.
- Define propTypes for PatientDataPrintDialog (api, patientId,
  chartPrefs, onClose) and the sibling ClinicPatientsPrintDialog,
  matching the project's PropTypes conventions and existing
  defaults.
- Move the ExportDialog test out of the legacy test/ tree into
  __tests__/unit/app/components, folding its cases into the
  existing suite. Convert click interactions to userEvent and
  rename cases to the "should X when Y" pattern.
